cleaning the code
